+++ packages/implementations/web-components/src/input/Button/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The Button component allows us to initiate actions, make state or page changes.
 
+## 2024-08-06 `1.0.0`
+
+#### 🎉 New features
+
+- Added `appearance`, `as`, `full`, `size` and `variant` properties to the Component. ([#87](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/87) by [@junior](https://git.rarolabs.com.br/junior))
+- Added stories on Component. ([#87](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/87) by [@junior](https://git.rarolabs.com.br/junior))
+
 <!-- #### 🛠 Breaking changes -->
 
 <!-- #### 📚 3rd party library updates -->